third_party/WebKit/LayoutTests/inspector/device-emulation/device-emulation-980-only-viewport.html

<html>
<head>

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="device-emulation-test.js"></script>

<script>
// This test is based on http://jsbin.com/urowoh/latest.
setMetaViewport();
</script>

<style>
html {
    overflow-x: hidden;
}

body {
    margin: 0;
    min-height: 1000px;
    overflow-x: hidden;
}
</style>

<script>
function test()
{
    InspectorTest.testDeviceEmulation("device-emulation-980-only-viewport.html", 0, 0, 0, "w=980");
}
</script>

</head>
<body onload="runTest()">
<p>
Tests that device emulation affects media rules, viewport meta tag, body dimensions and window.screen.
</p>
</body>
</html>
