third_party/WebKit/LayoutTests/inspector/device-emulation/device-emulation-980-only-viewport-expected.txt

Tests that device emulation affects media rules, viewport meta tag, body dimensions and window.screen.

Page reloaded.
Emulating device: 0x0x0 viewport='w=980'
Emulation warning: You might need to reload the page for proper user agent spoofing and viewport rendering.
Device:
window.screenX = 0px
window.screenY = 0px
Viewport: = ?w=980
@media orientation = landscape
window.orientation
Widths:
doc.docElem.clientWidth = 980px
doc.docElem.offsetWidth = 980px
doc.docElem.scrollWidth = 980px
doc.body.offsetWidth = 980px
doc.body.scrollWidth = 980px
Heights:
doc.docElem.clientHeight = 1000px
doc.docElem.offsetHeight = 1000px
doc.docElem.scrollHeight = 1000px
doc.body.offsetHeight = 1000px
doc.body.scrollHeight = 1000px
